Code is slightly better
